// src/lib/scripts/services/export/excel/services-export-excel.interface.d.ts
export declare type TExcelFileName = string | null;
export declare type TExcelExport = 'Single' | 'Multiple';
export interface IExcelResult {
    code: number;
    message: string;
}
export interface IExcelExportSingleArray {
    userData: any[];
    sheetName: string;
    fileName: string;
}
export interface IExcelExportMultipleArray {
    data: IArrayMatrixToExcel[];
    fileName?: TExcelFileName;
}
export interface IExcelJsonToFileProps {
    Title: string;
    Author: string;
    Company: string;
    Category: string;
}
export interface IArrayMatrixToExcel {
    content: any[];
    sheetName: string;
}